Use BadgedEx from ControlzEx   Update the other dependencies too
