+++ MaterialDesignThemes.Wpf/paket.references
@@ -1,2 +1,3 @@
 File:PackIconBase.cs ControlzEx
 File:PopupEx.cs ControlzEx
+File:BadgedEx.cs ControlzEx
